+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 This is intended as an overview of the major changes
 
-v4.6
+v5.0
 ===
 - Bump Flink to 1.5 and test both DataSet and DataStream use.
 - Bump Apache Drill version to 1.13